Move query router handlers to their own file
